Move transactionsList to bridgeForm
